Move ScopedStreamFormat to its own file
